11.jan up

+++ database/migrations/2020_01_11_091644_add_opinion_id_to_user_table.php
@@ -14,7 +14,7 @@ class AddOpinionIdToUserTable extends Migration
     public function up()
     {
         Schema::table('user', function (Blueprint $table) {
-            $table->BIGINT('opinion_id')->after('password');
+            $table->int('opinion_id')->after('password');
         });
     }
